+++ src/components/TodoHeader.jsx
@@ -1,0 +1,5 @@
+function TodoHeader() {
+    return <h1>TODO 앱</h1>
+}
+
+export default TodoHeader;
